fixed deployment script error checking

+++ deploy.sh
@@ -5,10 +5,15 @@ DATE=`date "+%Y-%m-%dT%H_%M_%S"`
 function cpsafe() {
  SRC=$1
  DST=$2
- echo "Copying '$SRC' to $DST"
- [[ -f "$SRC" ]] || echo "$SRC doesn't exist!" && return 1
- [[ -f "$DST" ]] && mv "$DST" "$DST.old-$DATE"
- cp "$SRC" "$DST"
+ printf "Copying '$SRC' to $DST..."
+ if [[ -f "$SRC" ]]; then 
+  [[ -f "$DST" ]] && mv "$DST" "$DST.old-$DATE"
+  cp "$SRC" "$DST"
+  printf " Copied.\n"
+ else
+  printf "\n *ERROR*: $SRC doesn't exist!\n\n"
+  return 1
+ fi
 }
 
 cpsafe ./.bashrc ~/.bashrc
